requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
aiofiles
langchain
typing-extensions
tqdm
requests
langchain-huggingface
transformers
huggingface-hub
aiohttp
